update front-end Admin

+++ ShopApp-BE/src/main/java/com/example/ShopApp_BE/DTO/OrderDetailDTO.java
@@ -13,7 +13,7 @@ public class OrderDetailDTO{
 
     private String name;
 
-    private Integer numberOfProducts;
+    private Long numberOfProducts;
 
     private Double price;
